% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Text_Proc}
\alias{Text_Proc}
\title{Preprocessing for textual data}
\usage{
Text_Proc(text, rem_num = F, rem_spec = F, lemma = F,
  lib_stop = c("snowball", "stopwords-iso", "smart"), add_stop = NULL,
  ngram_min = 1, ngram_max = 1, term_min = 1, term_max = Inf,
  doc_prop_min = 0, doc_prop_max = 1, doc_cnt_min = 1,
  doc_cnt_max = Inf, voc_cnt_max = Inf, write_mat = F,
  get_wordcloud = F, wc_term_max = Inf, wc_col = "Dark2")
}
\arguments{
\item{text}{A character vector with the textual data (words/strings).}

\item{rem_num}{Logical specifing if digits in the text should be removed. FALSE by default.}

\item{rem_spec}{Logical specifying if special characters (other than the 14 commonly used punctuation marks) such as @, #, & etc should be removed. FALSE by default.}

\item{lemma}{Logical specifying if the words in the text should be lemmatized to their base forms. FALSE by default.}

\item{lib_stop}{Libraries to be used for stop word removal.}

\item{add_stop}{Character vector of user defined stop words. This is optional.}

\item{ngram_min}{The minimum number of words to be considered for ngrams.}

\item{ngram_max}{The maximum number of words to be considered for ngrams.}

\item{term_min}{The minimum number of times a word should be present in the text for it to be recognised as a keyword.}

\item{term_max}{The maximum number of times a word can be present in the text for it to be recognised as a keyword.}

\item{doc_prop_min}{Minimum proportion of documents which should contain the term.}

\item{doc_prop_max}{Maximum proportion of documents which should contain the term.}

\item{doc_cnt_min}{Minimum number of documents that should contain the term for it ro be considered as a keyword.}

\item{doc_cnt_max}{Maximum number of documents that can contain the term for it to be considered a keyword.}

\item{voc_cnt_max}{The maximum number of keywords to be included in the vocabulary. Keywords with higher frequencies are retained.}

\item{write_mat}{Logical specifying if the Document Term Matrix should be written out as a CSV file. FALSE by default.}

\item{get_wordcloud}{Logical specifying if a word cloud should be created. FALSE by default.}

\item{wc_term_max}{The maximum number of terms to be included in the word cloud. Terms with higher frequencies are retained.}

\item{wc_col}{The color scheme to be used for the word cloud. Color schemes from RColorBrewer are supported.}
}
\value{
A list containing the processed text, the vocabulary (Term Frequency Table) and the DTM.
}
\description{
The function allows users to clean textual data by removing punctuation marks, whitespaces and stop words. A vocabulary and Document Term Matrix are then generated.
Digits and other special characters can also be removed and words can be lemmatized to their base forms.
Users can also choose to create word clouds for visualization and save the DTM locally.
}
